refactoring of the precompiled contract
